+++ packages/hw-ledger-transports/src/ledgerhq-web.d.ts
@@ -1,0 +1,8 @@
+// Copyright 2017-2021 @polkadot/hw-ledger authors & contributors
+// SPDX-License-Identifier: Apache-2.0
+
+declare module '@ledgerhq/hw-transport-webhid' {
+  import Transport from '@ledgerhq/hw-transport';
+
+  export default class TransportWebHid extends Transport {}
+}